Move ED tests
